setup mocha and chai for unit testing
